mock-api/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.3
pydantic==2.9.2
strawberry-graphql[fastapi]==0.247.2
